requirements.txt
astroid==2.1.0
Django==2.1.4
django-appconf==1.0.2
django-autocomplete-light==3.3.2
django-js-reverse==0.8.2
django-select2==6.3.1
djangoajax==3.0.3
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2-binary==2.7.6.1
pylint==2.2.2
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
wrapt==1.10.11
